+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/contracts/application/service/ContractServiceReadOnly.java
@@ -18,19 +18,10 @@
  ********************************************************************************/
 package org.eclipse.tractusx.traceability.contracts.application.service;
 
-import org.eclipse.tractusx.irs.edc.client.contract.model.exception.ContractAgreementException;
 import org.eclipse.tractusx.traceability.common.model.PageResult;
 import org.eclipse.tractusx.traceability.common.request.PageableFilterRequest;
 import org.eclipse.tractusx.traceability.contracts.domain.model.Contract;
-import org.eclipse.tractusx.traceability.contracts.domain.model.ContractAgreement;
-import org.eclipse.tractusx.traceability.contracts.domain.model.ContractType;
 
-import java.util.List;
-
-public interface ContractService {
-
-    void saveContractAgreements(List<String> contractAgreementIds, ContractType contractType) throws ContractAgreementException;
-
-    void saveAll(List<ContractAgreement> contractAgreements);
-
+public interface ContractServiceReadOnly {
+    PageResult<Contract> getContracts(PageableFilterRequest pageableFilterRequest);
 }
