+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/contracts/domain/repository/ContractRepositoryReadOnly.java
@@ -16,21 +16,14 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
-package org.eclipse.tractusx.traceability.contracts.application.service;
+package org.eclipse.tractusx.traceability.contracts.domain.repository;
 
-import org.eclipse.tractusx.irs.edc.client.contract.model.exception.ContractAgreementException;
 import org.eclipse.tractusx.traceability.common.model.PageResult;
-import org.eclipse.tractusx.traceability.common.request.PageableFilterRequest;
+import org.eclipse.tractusx.traceability.common.model.SearchCriteria;
 import org.eclipse.tractusx.traceability.contracts.domain.model.Contract;
-import org.eclipse.tractusx.traceability.contracts.domain.model.ContractAgreement;
-import org.eclipse.tractusx.traceability.contracts.domain.model.ContractType;
+import org.springframework.data.domain.Pageable;
 
-import java.util.List;
+public interface ContractRepositoryReadOnly<T> {
 
-public interface ContractService {
-
-    void saveContractAgreements(List<String> contractAgreementIds, ContractType contractType) throws ContractAgreementException;
-
-    void saveAll(List<ContractAgreement> contractAgreements);
-
+    PageResult<Contract> getContractsByPageable(Pageable pageable, SearchCriteria searchCriteria);
 }
