+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/contracts/infrastructure/repository/ContractAsPlannedRepositoryImpl.java
@@ -18,76 +18,27 @@
  ********************************************************************************/
 package org.eclipse.tractusx.traceability.contracts.infrastructure.repository;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import lombok.RequiredArgsConstructor;
+import lombok.experimental.SuperBuilder;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.tuple.ImmutablePair;
-import org.eclipse.tractusx.irs.edc.client.contract.model.EdcContractAgreementNegotiationResponse;
-import org.eclipse.tractusx.irs.edc.client.contract.model.EdcContractAgreementsResponse;
 import org.eclipse.tractusx.irs.edc.client.contract.model.exception.ContractAgreementException;
 import org.eclipse.tractusx.irs.edc.client.contract.service.EdcContractAgreementService;
-import org.eclipse.tractusx.traceability.common.model.PageResult;
-import org.eclipse.tractusx.traceability.common.model.SearchCriteria;
-import org.eclipse.tractusx.traceability.common.repository.BaseSpecification;
-import org.eclipse.tractusx.traceability.contracts.domain.exception.ContractException;
 import org.eclipse.tractusx.traceability.contracts.domain.model.Contract;
-import org.eclipse.tractusx.traceability.contracts.domain.model.ContractAgreement;
 import org.eclipse.tractusx.traceability.contracts.domain.model.ContractType;
 import org.eclipse.tractusx.traceability.contracts.domain.repository.ContractRepository;
 import org.eclipse.tractusx.traceability.contracts.infrastructure.model.ContractAgreementAsPlannedEntity;
 import org.eclipse.tractusx.traceability.contracts.infrastructure.model.ContractAgreementBaseEntity;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Component;
 
-import java.time.Instant;
-import java.time.OffsetDateTime;
-import java.time.ZoneId;
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
-import java.util.Map;
-import java.util.Objects;
 import java.util.stream.Collectors;
 
-import static org.apache.commons.collections4.ListUtils.emptyIfNull;
-
 @Component
-@RequiredArgsConstructor
+@SuperBuilder
 @Slf4j
-public class ContractAsPlannedRepositoryImpl implements ContractRepository<ContractAgreementAsPlannedEntity> {
+public class ContractAsPlannedRepositoryImpl extends ContractRepositoryImplBase implements ContractRepository<ContractAgreementAsPlannedEntity> {
 
     private final EdcContractAgreementService edcContractAgreementService;
     private final JpaContractAgreementAsPlannedRepository contractAgreementRepository;
-    private final ObjectMapper objectMapper;
-
-    @Override
-    public PageResult<Contract> getContractsByPageable(Pageable pageable, SearchCriteria searchCriteria) {
-        try {
-            List<ContractAgreementAsPlannedSpecification> contractAgreementSpecifications = emptyIfNull(searchCriteria.getSearchCriteriaFilterList()).stream()
-                    .map(ContractAgreementAsPlannedSpecification::new)
-                    .toList();
-            Specification<ContractAgreementAsPlannedEntity> specification = BaseSpecification.toSpecification(contractAgreementSpecifications);
-            Page<ContractAgreementAsPlannedEntity> contractAgreementEntities = contractAgreementRepository.findAll(specification, pageable);
-
-            if (contractAgreementEntities.getContent().isEmpty()) {
-                log.warn("Cannot find contract agreement Ids for asset ids in searchCriteria: " + searchCriteria.getSearchCriteriaFilterList());
-                return new PageResult<>(List.of(), 0, 0, 0, 0L);
-            }
-
-            return new PageResult<>(fetchEdcContractAgreements(contractAgreementEntities.getContent()),
-                    contractAgreementEntities.getPageable().getPageNumber(),
-                    contractAgreementEntities.getTotalPages(),
-                    contractAgreementEntities.getPageable().getPageSize(),
-                    contractAgreementEntities.getTotalElements());
-
-        } catch (ContractAgreementException e) {
-            throw new ContractException(e);
-        }
-
-    }
 
     @Override
     public void saveAllContractAgreements(List<String> contractAgreementIds, ContractType contractType) throws ContractAgreementException {
@@ -99,7 +50,13 @@ public class ContractAsPlannedRepositoryImpl implements ContractRepository<Contr
                         .build())
                 .collect(Collectors.toList());
 
-        List<Contract> contracts = fetchEdcContractAgreements(contractAgreementEntities);
+        List<ContractAgreementBaseEntity> baseEntities = contractAgreementEntities
+                .stream()
+                .map(entity -> (ContractAgreementBaseEntity) entity)
+                .toList();
+
+
+        List<Contract> contracts = fetchEdcContractAgreements(baseEntities);
         List<ContractAgreementAsPlannedEntity> contractAgreementsUpdated = ContractAgreementAsPlannedEntity.fromList(contracts, contractType);
         contractAgreementRepository.saveAll(contractAgreementsUpdated);
     }
@@ -107,77 +64,6 @@ public class ContractAsPlannedRepositoryImpl implements ContractRepository<Contr
     @Override
     public void saveAll(List<ContractAgreementAsPlannedEntity> contractAgreements) {
         contractAgreementRepository.saveAll(contractAgreements);
-    }
-
-    @Override
-    public void save(ContractAgreement contractAgreement) {
-        contractAgreementRepository.save(ContractAgreementAsPlannedEntity.fromDomainToEntity(contractAgreement));
-    }
-
-    @Override
-    public List<ContractAgreementAsPlannedEntity> findAll() {
-        return List.of();
-    }
-
-    private List<Contract> fetchEdcContractAgreements(List<ContractAgreementAsPlannedEntity> contractAgreementEntities) throws ContractAgreementException {
-        List<String> contractAgreementIds = contractAgreementEntities.stream().filter(Objects::nonNull).map(ContractAgreementAsPlannedEntity::getContractAgreementId).toList();
-        log.info("Trying to fetch contractAgreementIds from EDC: " + contractAgreementIds);
-
-        List<EdcContractAgreementsResponse> contractAgreements = edcContractAgreementService.getContractAgreements(contractAgreementIds);
-
-        validateContractAgreements(contractAgreementIds, contractAgreements);
-
-        Map<String, ContractType> contractTypes = contractAgreementEntities.stream()
-                .collect(Collectors.toMap(ContractAgreementAsPlannedEntity::getContractAgreementId, ContractAgreementAsPlannedEntity::getType));
-
-        Map<String, EdcContractAgreementNegotiationResponse> contractNegotiations = contractAgreements.stream()
-                .map(agreement -> new ImmutablePair<>(agreement.contractAgreementId(),
-                        edcContractAgreementService.getContractAgreementNegotiation(agreement.contractAgreementId()))
-                ).collect(Collectors.toMap(ImmutablePair::getLeft, ImmutablePair::getRight));
-
-
-        return contractAgreements.stream().map(contractAgreement ->
-                {
-                    try {
-                        String globalAssetId = contractAgreementEntities.stream()
-                                .filter(contractAgreementViewEntity -> contractAgreementViewEntity.getContractAgreementId().equals(contractAgreement.contractAgreementId()))
-                                .findFirst()
-                                .map(ContractAgreementBaseEntity::getGlobalAssetId)
-                                .orElse(null);
-                        return Contract.builder()
-                                .globalAssetId(globalAssetId)
-                                .contractId(contractAgreement.contractAgreementId())
-                                .counterpartyAddress(contractNegotiations.get(contractAgreement.contractAgreementId()).counterPartyAddress())
-                                .creationDate(OffsetDateTime.ofInstant(Instant.ofEpochSecond(contractAgreement.contractSigningDate()), ZoneId.systemDefault()))
-                                .state(contractNegotiations.get(contractAgreement.contractAgreementId()).state())
-                                .policy(objectMapper.writeValueAsString(contractAgreement.policy()))
-                                .type(contractTypes.get(contractAgreement.contractAgreementId()))
-                                .build();
-                    } catch (JsonProcessingException e) {
-                        throw new RuntimeException(e);
-                    }
-                }
-        ).toList();
-    }
-
-    private void validateContractAgreements(List<String> contractAgreementIds, List<EdcContractAgreementsResponse> contractAgreements) {
-        ArrayList<String> givenList = new ArrayList<>(contractAgreementIds);
-        Collections.sort(givenList);
-
-        List<String> expectedList = contractAgreements.stream()
-                .map(EdcContractAgreementsResponse::contractAgreementId)
-                .sorted()
-                .toList();
-        log.info("EDC responded with the following contractAgreementIds: " + expectedList);
-
-        // Filter the givenList to find out which IDs are missing in the expectedList
-        List<String> missingIds = givenList.stream()
-                .filter(id -> !expectedList.contains(id))
-                .toList();
-
-        if (!missingIds.isEmpty()) {
-            log.warn("Cannot find the following contract agreement IDs in EDC: " + missingIds);
-        }
     }
 
 }
